+++ RogueLikeMADY/src/main/java/com/mady/utils/entities/factories/items/PotionVie.java
@@ -16,10 +16,10 @@ public class PotionVie extends AbstractItem {
     @Override
     public void act(Player player) {
 // une potion de vie peut faire gagner des points de vie bonus supérieur aux max hitpoint le comportement 101/100 est voulu pour permettre de stack de la vie en revanche le joueur ne pourra pas se reposer pour avoir plus de  maxhitpoint
-        int oldHp = (int) player.getHitPoints();
+        int oldHp = player.getHitPoints();
         player.setHitPoints(player.getHitPoints() + getDamages());
         Util.currentAction.append(Ansi.colorize(String.format("%s vous augmente la vie de %d à %d (+%d)\n",
-                getName(), oldHp, (int) player.getHitPoints(), getDamages()), Attribute.GREEN_TEXT()));
+                getName(), oldHp, player.getHitPoints(), getDamages()), Attribute.GREEN_TEXT()));
 
     }
 }
